Add files via upload
This is the first module self assessment that I had to complete for the self pace study course. I had to write a program that will calculate the cost of a custom cup of coffee at a gourmet coffee shop, based on the size of  the cup, the type of coffee selected, and flavors that can be added to the coffee.